moving component to separate directory
