+++ src/app/app.module.ts
@@ -5,13 +5,13 @@ import { RouterModule } from '@angular/router';
 import { ApolloModule } from 'angular2-apollo';
 import { TimeAgoPipe } from 'angular2-moment';
 
-import { AppComponent } from './app.component';
-import { NavigationComponent } from './navigation.component';
-import { ProfileComponent } from './profile.component';
-import { NewEntryComponent } from './new-entry.component';
-import { FEED_DECLARATIONS } from './feed';
-import { COMMENTS_DECLARATIONS } from './comments';
-import { SHARED_DECLARATIONS } from './shared';
+import { AppComponent } from './components/app.component';
+import { NavigationComponent } from './components/navigation.component';
+import { ProfileComponent } from './components/profile.component';
+import { NewEntryComponent } from './components/new-entry.component';
+import { FEED_DECLARATIONS } from './components/feed';
+import { COMMENTS_DECLARATIONS } from './components/comments';
+import { SHARED_DECLARATIONS } from './components/shared';
 import { routes } from './routes';
 import { client } from './client';
