+++ src/app/components/app.component.ts
@@ -2,6 +2,6 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'app-root',
-  templateUrl: './app.component.html'
+  templateUrl: 'app.component.html'
 })
 export class AppComponent {}
