+++ src/app/components/feed/index.ts
