+++ src/app/components/feed/vote-buttons.component.html
