+++ src/app/components/navigation.component.ts
@@ -3,7 +3,7 @@ import { Router } from '@angular/router';
 
 @Component({
   selector: 'navigation',
-  templateUrl: './navigation.component.html'
+  templateUrl: 'navigation.component.html'
 })
 export class NavigationComponent {
   constructor(public router: Router) {}
