+++ src/app/components/new-entry.component.html
