+++ src/app/components/profile.component.ts
@@ -15,7 +15,7 @@ const CurrentUserQuery = gql`
 
 @Component({
   selector: 'profile',
-  templateUrl: './profile.component.html'
+  templateUrl: 'profile.component.html'
 })
 export class ProfileComponent implements OnInit, OnDestroy {
   loading: boolean = true;
