+++ src/app/components/shared/loading.component.ts
