--- src/app/index.ts
@@ -1,2 +1,0 @@
-export * from './app.component';
-export * from './app.module';
